Refactor modal

+++ vue/cukcuk/src/components/Table.vue
@@ -13,7 +13,7 @@
         </button>
       </div>
       <div class="right">
-        <button class="btn-add" @click="showModal('add')">
+        <button class="btn-add" @click="openModal('add', {})">
           <span class="icon"></span>
           <span> Thêm mới </span>
         </button>
@@ -65,7 +65,7 @@
               <div class="table__tooltip">
                 <button
                   class="table__tooltip--icon1"
-                  @click="showModal('edit', item)"
+                  @click="openModal('edit', item)"
                 >
                   <span class="material-symbols-outlined"> edit </span>
                 </button>
@@ -94,80 +94,23 @@
         </span>
       </div>
     </div>
-
-    <Modal
-      v-if="isModalVisible"
-      @closeFromTable="closeModal"
-      :action="modalAction"
-      :item="modalItem"
-    ></Modal>
-    <Dialog
-      v-if="isShowDiaglog"
-      :action="dialogAction"
-      :title="dialogTitle"
-      :description="dialogDesc"
-      :type="dialogType"
-      @closeDialog="closeDialog"
-      @delete="deleteEmployeeDB"
-    ></Dialog>
+    <Dialog ref="dialog" v-model="modalItem"></Dialog>
   </div>
 </template>
 
 <script>
-import Modal from '../components/Modal.vue';
 import Dialog from '../components/Dialog.vue';
 
 /* eslint-disable */
 export default {
   name: 'Table',
-  props: {},
   data() {
     return {
       items: [],
       checkedItems: [],
-      isModalVisible: false,
-      modalAction: '',
-      modalItem: {},
-
-      dialogTitle: '',
-      dialogAction: '',
-      dialogDesc: '',
-      dialogType: '',
-      isShowDiaglog: false,
-      isDialogConfirm: false,
     };
   },
   methods: {
-    showModal(action, item) {
-      this.modalAction = action;
-      if (action == 'edit') {
-        this.modalItem = item;
-      }
-
-      if (action == 'add') {
-        this.modalItem = {};
-      }
-
-      this.isModalVisible = true;
-    },
-    showDialog(action, title, desc, type) {
-      this.dialogAction = action;
-      this.dialogTitle = title;
-      this.dialogDesc = desc;
-      this.dialogType = type;
-      this.isShowDiaglog = true;
-    },
-    closeModal() {
-      this.isModalVisible = false;
-    },
-    closeDialog() {
-      this.isShowDiaglog = false;
-    },
-    changeStateDialog(state) {
-      console.log(state);
-      this.isDialogConfirm = state;
-      console.log(this.isDialogConfirm);
-    },
     convertGender(x) {
       if (x == 0) return 'Nữ';
       else return 'Nam';
@@ -214,55 +157,60 @@ export default {
         console.error(error);
       }
     },
+    openModal(action, item) {
+      this.$emitter.emit('showModal', action, item);
+    },
     async deleteEmployee() {
       if (this.checkedItems.length == 0) {
-        // alert('Không có item nào được chọn!');
-        this.showDialog('0', 'Thông báo', 'Không có item nào được chọn!', 1);
+        this.$refs.dialog.show({
+          title: 'Thông báo',
+          desc: 'Không có item nào được chọn!',
+          type: 1,
+        });
       } else {
-        this.showDialog(
-          'delete',
-          'Thông báo',
-          'Bạn có chắc chắn muốn xóa không?',
-          2
-        );
-      }
-    },
-    async deleteEmployeeDB() {
-      try {
-        for (let id of this.checkedItems) {
-          await fetch(`${process.env.ENDPOINT}/employees/${id}`, {
-            method: 'DELETE',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-          });
-          this.items = this.items.filter((item) => item.EmployeeId != id);
-          this.checkedItems = this.checkedItems.filter(
-            (checkedId) => checkedId !== id
-          );
-        }
-        this.showDialog('0', 'Thông báo', 'Xóa thành công', 1);
+        const ok = await this.$refs.dialog.show({
+          title: 'Thông báo',
+          desc: 'Bạn có chắc chắn muốn xóa không?',
+          type: 2,
+        });
+        if (ok) {
+          try {
+            for (let id of this.checkedItems) {
+              await fetch(`${process.env.ENDPOINT}/employees/${id}`, {
+                method: 'DELETE',
+                headers: {
+                  'Content-Type': 'application/json',
+                },
+              });
+              this.items = this.items.filter((item) => item.EmployeeId != id);
+              this.checkedItems = this.checkedItems.filter(
+                (checkedId) => checkedId !== id
+              );
+            }
+            this.$refs.dialog.show({
+              title: 'Thông báo',
+              desc: 'Xóa thành công',
+              type: 1,
+            });
 
-        // TODO: fix this
-        // window.location.reload();
-      } catch (e) {
-        console.error(e);
-        throw e;
+            // TODO: fix this
+            // window.location.reload();
+          } catch (e) {
+            console.error(e);
+            throw e;
+          }
+        }
       }
     },
   },
   components: {
-    Modal,
     Dialog,
   },
   mounted() {
     try {
-      console.log(process.env.ENDPOINT);
       fetch(`${process.env.ENDPOINT}/employees`)
         .then((res) => res.json())
         .then((data) => {
-          console.log(data);
-          // console.log(data[0]);
           this.items = data;
         });
     } catch (error) {
